sogno/coordinator/requirements.txt
paho-mqtt
datetime
typing
requests
pandas